+++ History.md
@@ -2,6 +2,7 @@
 ===================
 
   * v0.11.x process.nextTick compatibility [#5](https://github.com/fivdi/lcd/issues/5)
+  * Example print-twice-20x4.js added
   * Updated dependency: onoff 0.3.0 -> 0.3.1
   * Updated dependency: q 0.9.7 -> 1.0.1